requirements.txt
streamlit==1.39.0
numpy
pillow
rdkit-pypi
selfies
torch==2.5.1
torch-geometric==2.6.1
protobuf==3.20.3
